fix: skip deployment script generation in Docker mode
In Docker mode (pip install), users don't have the full repository
locally, so we can't generate deployment scripts on the host machine.

The realm data and source code are successfully generated in the Docker
container. Users will need to use the Docker image or clone the full
repository to deploy the realm.

This completes the fix for test-cli-docker-mode.